+++ FKNI.Infraestructure/Models/Promociones.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
@@ -9,7 +8,6 @@ namespace FKNI.Infraestructure.Models;
 public partial class Promociones
 {
     public int IdPromocion { get; set; }
-
     [Required(ErrorMessage = "Debe seleccionar un tipo de promoción")]
     public string? TipoPromocion { get; set; }
 
@@ -17,7 +15,7 @@ public partial class Promociones
 
     public int? IdCategoria { get; set; }
 
-    public decimal Descuento { get; set; }
+    public double Descuento { get; set; }
 
     public DateTime? FechaInicio { get; set; }
 
@@ -30,4 +28,3 @@ public partial class Promociones
     [NotMapped]
     public string BuscarProducto { get; set; } = null!;
 }
-
